+++ lib/foundation/routing/app_router.dart
@@ -1,8 +1,6 @@
 import 'package:auto_route/auto_route.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:tictactoe/core/providers/services/user.service.provider.dart';
-import 'package:tictactoe/presentation/screens/authentication/authentication.screen.dart';
 import 'package:tictactoe/presentation/screens/home/home.screen.dart';
+import 'package:tictactoe/presentation/screens/lobby/lobby.screen.dart';
 
 part 'app_router.gr.dart';
 
@@ -14,38 +12,7 @@ class AppRouter extends RootStackRouter {
 
   @override
   List<AutoRoute> get routes => <AutoRoute>[
-    AutoRoute(page: AuthenticationRoute.page),
-    AutoRoute(
-      page: HomeRoute.page,
-      initial: true,
-      guards: <AutoRouteGuard>[AuthGuard()],
-    ),
+    AutoRoute(page: LobbyRoute.page, initial: true),
+    AutoRoute(page: HomeRoute.page),
   ];
-}
-
-/// Auth guard
-class AuthGuard extends AutoRouteGuard {
-  @override
-  Future<void> onNavigation(
-    NavigationResolver resolver,
-    StackRouter router,
-  ) async {
-    final ProviderContainer container = ProviderScope.containerOf(
-      resolver.context,
-      listen: false,
-    );
-    await container.read(
-      userServiceProvider.future,
-    );
-
-    resolver.next();
-
-    // if (userService.isAuthenticated) {
-    //   debugPrint('guard - isAuthenticated');
-    //   resolver.next();
-    // } else {
-    //   debugPrint('guard - notAuthenticated');
-    //   await resolver.redirectUntil(const AuthenticationRoute());
-    // }
-  }
 }